refactor[#453] : 잘못 선언된 lib 위치 변경
