+++ src/entities/marking/api/getBoundaryMarkerList.ts
@@ -1,6 +1,6 @@
 import { skipToken, useQuery } from "@tanstack/react-query";
 import { useMapStore } from "@/features/map/store";
-import { useKMeansClustering } from "@/entities/map/ui/lib";
+import { useKMeansClustering } from "@/entities/map/lib";
 import { apiClient } from "@/shared/lib";
 import { useAuthStore } from "@/shared/store";
 import { MARKER_END_POINT } from "../constants";
